rename correlation as pearsonr
