+++ src/scores/continuous/correlation/__init__.py
@@ -1,0 +1,7 @@
+"""
+Import the functions from the implementations into the public API
+"""
+
+from scores.continuous.correlation.correlation_impl import pearsonr
+
+__all__ = ["pearsonr"]
